Rename gradients.py to vanilla_gradients.py
