+++ tf_explain/callbacks/__init__.py
@@ -5,7 +5,7 @@ This module regroups all the methods as tf.keras.Callback.
 """
 from .activations_visualization import ActivationsVisualizationCallback
 from .grad_cam import GradCAMCallback
-from .gradients import VanillaGradientsCallback
+from .vanilla_gradients import VanillaGradientsCallback
 from .integrated_gradients import IntegratedGradientsCallback
 from .occlusion_sensitivity import OcclusionSensitivityCallback
 from .smoothgrad import SmoothGradCallback
